Moved env type enum to Common
