+++ src/Common/EnvironmentType.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT license.
 // --------------------------------------------------------------------------------------------
 
-namespace Microsoft.Oryx.BuildScriptGenerator
+namespace Microsoft.Oryx.Common
 {
     public enum EnvironmentType
     {
